+++ lib/validators/modelValidator.ts
@@ -1,5 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
+import * as amd from "@azure/openapi-markdown"
 import { filter, toArray } from "@ts-common/iterator"
 import { JsonRef } from "@ts-common/json"
 import * as jsonParser from "@ts-common/json-parser"
@@ -784,9 +785,14 @@ export class ModelValidator extends SpecValidator<SpecValidationResult> {
             foundIssues = true
             break
           }
-          // replacing forward slashes with empty string because this messes up Sways regex
+
+          // replacing characters that may cause validator failed  with empty string because this messes up Sways regex
           // validation of path segment.
-          parameterValue = parameterValue.replace(/\//gi, "")
+          if (!utils.isUrlEncoded(parameterValue as string)) {
+            // TODO: we can get the scheme from parameterValue if the useSchemePrefix is setting false in the x-ms-parameterized-host,
+            // then check if it can match to the swagger scheme.
+            parameterValue = parameterValue.replace(/[^0-9a-zA-Z._]/gi, "")
+          }
         }
         const paramType = location + "Parameters"
         if (!options[paramType]) {
@@ -901,23 +907,48 @@ export class ModelValidator extends SpecValidator<SpecValidationResult> {
       if (!!requestValidationErrors && requestValidationErrors.length > 0) {
         if (exampleFilePath !== undefined) {
           requestValidationErrors.forEach(error => {
-            const position = getDescendantFilePosition(
+            const positionInfo = getDescendantFilePosition(
               this.exampleJsonMap.get(exampleFilePath) as JsonRef,
               error.path
             )
-            error.title = `{"path":["${error.path}"], "position":"${JSON.stringify(
-              position
-            )}", "url":"${exampleFilePath}"}`
+            if (!error.path) {
+              error.path = ""
+            }
+            const titleObj: any = {
+              path: Array.isArray(error.path) ? [...error.path] : [error.path],
+              position: positionInfo,
+              url: exampleFilePath
+            }
+            error.title = JSON.stringify(titleObj)
           })
+          // process suppression for request validation errors
+          if (this.suppression) {
+            const requestParameterSuppressions = this.suppression.directive.filter(
+              item => item.suppress === "INVALID_REQUEST_PARAMETER"
+            )
+            if (requestParameterSuppressions && requestParameterSuppressions.length > 0) {
+              requestValidationErrors = this.applySuppression(
+                requestValidationErrors,
+                requestParameterSuppressions
+              )
+            }
+          }
         }
       }
-      const e = this.constructErrorObject({
-        code: C.ErrorCodes.RequestValidationError,
-        message: msg,
-        innerErrors: requestValidationErrors
-      })
-      operationResult.request.error = e
-      log.error(`${msg}:\n`, e)
+      if (requestValidationErrors && requestValidationErrors.length > 0) {
+        const e = this.constructErrorObject({
+          code: C.ErrorCodes.RequestValidationError,
+          message: msg,
+          innerErrors: requestValidationErrors
+        })
+        operationResult.request.error = e
+        log.error(`${msg}:\n`, e)
+      } else {
+        msg = "Request parameters is valid."
+        operationResult.request.isValid = true
+        operationResult.request.result = msg
+        log.info(`${msg}`)
+      }
     } else if (requestValidationWarnings) {
       operationResult.request.warning = requestValidationWarnings
       log.debug(`${msg}:\n`, requestValidationWarnings)
@@ -926,6 +957,31 @@ export class ModelValidator extends SpecValidator<SpecValidationResult> {
       operationResult.request.result = msg
       log.info(`${msg}`)
     }
+  }
+
+  private applySuppression(
+    errors: ModelValidationError[],
+    suppressionItems: amd.SuppressionItem[]
+  ): ModelValidationError[] {
+    const notSuppressedErrors: ModelValidationError[] = []
+    errors.forEach(item => {
+      if (!item.message || !this.existSuppression(suppressionItems, item.message)) {
+        notSuppressedErrors.push(item)
+      }
+    })
+    return notSuppressedErrors
+  }
+
+  private existSuppression(suppressionItems: amd.SuppressionItem[], message: string): boolean {
+    for (const item of suppressionItems) {
+      if (item["text-matches"] !== undefined) {
+        const regex = new RegExp(item["text-matches"])
+        if (regex.test(message)) {
+          return true
+        }
+      }
+    }
+    return false
   }
 
   private constructResponseResult(