+++ test/liveValidatorTests.ts
@@ -10,7 +10,7 @@ import { ResponsesObject } from "yasway"
 import * as Constants from "../lib/util/constants"
 import { LiveValidator } from "../lib/validators/liveValidator"
 
-const numberOfSpecs = 10
+const numberOfSpecs = 11
 jest.setTimeout(150000)
 
 describe("Live Validator", () => {
@@ -215,10 +215,10 @@ describe("Live Validator", () => {
       }
       // 'microsoft.unknown' -> 'unknown-api-version'
       assert.strictEqual(4, p[Constants.unknownApiVersion].post.length)
-      assert.strictEqual(11, p[Constants.unknownApiVersion].get.length)
+      assert.strictEqual(13, p[Constants.unknownApiVersion].get.length)
       assert.strictEqual(3, p[Constants.unknownApiVersion].head.length)
-      assert.strictEqual(5, p[Constants.unknownApiVersion].put.length)
-      assert.strictEqual(5, p[Constants.unknownApiVersion].delete.length)
+      assert.strictEqual(6, p[Constants.unknownApiVersion].put.length)
+      assert.strictEqual(6, p[Constants.unknownApiVersion].delete.length)
       assert.strictEqual(1, p[Constants.unknownApiVersion].patch.length)
     })
     it("should initialize for batch", async () => {
@@ -341,6 +341,50 @@ describe("Live Validator", () => {
   })
 
   describe("Initialize cache and search", () => {
+    it("should return zero result when search for unknown method in unknown RP unknown apiversion operations", async () => {
+      const options = {
+        directory: "./test/liveValidation/swaggers/specification/unknown-rp",
+        swaggerPathsPattern: ["**/*.json"]
+      }
+      const requestUrl =
+        "https://management.azure.com/" +
+        "subscriptions/randomSub/resourceGroups/randomRG/providers/providers/Microsoft.Storage" +
+        "/3fa73e4b-d60d-43b2-a248-fb776fd0bf60" +
+        "?api-version=2018-09-01-preview"
+      const validator: any = new LiveValidator(options)
+      await validator.initialize()
+      // Operations to match is RoleAssignments_Create
+      const validationInfo = validator.parseValidationRequest(requestUrl, "Put", "randomId")
+      const operations = validator.getPotentialOperations(validationInfo).operations
+      assert.strictEqual(0, operations.length)
+    })
+    it("should fall back to return child operation in case of request url have parent and child resouces", async () => {
+      const options = {
+        directory: "./test/liveValidation/swaggers/specification/authorization",
+        swaggerPathsPattern: ["**/*.json"]
+      }
+      const requestUrl =
+        "https://management.azure.com/" +
+        "subscriptions/randomSub/resourceGroups/randomRG/providers/providers/Microsoft.Storage" +
+        "/storageAccounts/storageoy6qv/blobServices/default/containers" +
+        "/privatecontainer/providers/Microsoft.Authorization/roleAssignments" +
+        "/3fa73e4b-d60d-43b2-a248-fb776fd0bf60" +
+        "?api-version=2018-09-01-preview"
+      const validator: any = new LiveValidator(options)
+      await validator.initialize()
+      // Operations to match is RoleAssignments_Create
+      const validationInfo = validator.parseValidationRequest(requestUrl, "Put", "randomId")
+      const operations = validator.getPotentialOperations(validationInfo).operations
+      const pathObject = operations[0].pathObject
+      if (pathObject === undefined) {
+        throw new Error("pathObject is undefined")
+      }
+      assert.strictEqual(1, operations.length)
+      assert.strictEqual(
+        "/{scope}/providers/Microsoft.Authorization/roleAssignments/{roleAssignmentName}",
+        pathObject.path
+      )
+    })
     it("should return one matched operation for arm-storage", async () => {
       const options = {
         directory: "./test/liveValidation/swaggers/"
@@ -546,8 +590,8 @@ describe("Live Validator", () => {
       assert.strictEqual((errors[0] as any).code, "INVALID_RESPONSE_CODE")
     })
 
-    // should be case insensitive for api version
-    it("should be case-insensitive for resource provider and API version", async () => {
+    // should be case insensitive for paramter name and the value of api version, resource provider
+    it("should be case-insensitive for parameter name, resource provider and API version", async () => {
       const options = {
         directory:
           "./test/liveValidation/swaggers/specification/storage/resource-manager/Microsoft.Storage/2015-05-01-preview",
@@ -560,7 +604,7 @@ describe("Live Validator", () => {
       await validator.initialize()
       const result = validator.validateLiveRequestResponse({
         liveRequest: {
-          url: adjustedUrl,
+          url: adjustedUrl.toLocaleUpperCase(),
           method: "get",
           headers: {
             "content-type": "application/json"
@@ -708,7 +752,11 @@ describe("Live validator snapshot validation", () => {
     )
   })
 
-  test(`should return expected error for unresolvable reference`, async () => {
+  /**
+   * this case is invalid because we can detect unresolved reference erro in the stage of resolve spec
+   * TODO: this error code should be removed from the doc later
+   */
+  test.skip(`should return expected error for unresolvable reference`, async () => {
     const options = {
       directory: `${__dirname}/liveValidation/swaggers/`,
       isPathCaseSensitive: false,